refactor: StartAndGoalPosition 패키지 이동
- 해당 객체는 View에서 사용하는 Dto의 일종이므로 기존 board.position 보다는
view 패키지에 위치하는 것이 적절해보여 패키지를 이동했다.
